Break challenger primitives into files.
