+++ client/claimManager/submitClaim.js
